Merges WebhookCall and WebhookCallDetails
