+++ tests/Integration/Resource/WebhookCallTest.php
@@ -32,7 +32,18 @@ class WebhookCallTest extends TestCase
     {
         $webhookCall = (new ResourceBuilder())->build([
             'sys' => [
-                'type' => 'WebhookCallOverview',
+                'type' => 'WebhookCallDetails',
+            ],
+            'request' => [
+                'method' => 'POST',
+                'url' => 'https://www.example.com',
+                'headers' => [],
+                'body' => '{}',
+            ],
+            'response' => [
+                'statusCode' => 200,
+                'headers' => [],
+                'body' => '',
             ],
             'statusCode' => 200,
             'errors' => [],
@@ -42,7 +53,7 @@ class WebhookCallTest extends TestCase
             'responseAt' => '2016-03-01T08:43:22.330Z',
         ]);
 
-        $json = '{"sys":{"type":"WebhookCallOverview"},"statusCode":200,"errors":[],"eventType":"publish","url":"https:\/\/webhooks.example.com\/endpoint","requestAt":"2016-03-01T08:43:22.024Z","responseAt":"2016-03-01T08:43:22.330Z"}';
+        $json = '{"sys":{"type":"WebhookCallDetails"},"request":{"method":"POST","url":"https://www.example.com","headers":[],"body":"{}"},"response":{"statusCode":200,"url":"https://www.example.com","headers":[],"body":""},"statusCode":200,"errors":[],"eventType":"publish","url":"https:\/\/webhooks.example.com\/endpoint","requestAt":"2016-03-01T08:43:22.024Z","responseAt":"2016-03-01T08:43:22.330Z"}';
 
         $this->assertJsonStringEqualsJsonString($json, json_encode($webhookCall));
 
@@ -60,7 +71,7 @@ class WebhookCallTest extends TestCase
     {
         (new ResourceBuilder())
             ->build(['sys' => [
-                'type' => 'WebhookCallOverview',
+                'type' => 'WebhookCallDetails',
             ]], $webhookCall);
     }
 }
